WIP part 2
The main lesson learned today was to remember to include / before sources, such as img src=/images and not just write img src=images.

I attempted to get the slick slider working but that did not seem to work. The CSS parts are only aproximately 1/5 done, which means i'm behind schedule by 4/5 parts.